refactor(config): import/export global config
