+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "miniprogram-image",
-  "version": "1.9.0",
+  "version": "2.0.0",
   "description": "",
   "main": "miniprogram_dist/index.js",
   "types": "index.d.ts",
