--- test.ts
@@ -1,5 +1,0 @@
-const a:App.AppInstance={
-    __image_config:{
-        
-    }
-}
